+++ packages/components/components/elvis-modal/src/react/useLockBodyScroll.tsx
@@ -1,5 +1,6 @@
 import { useEffect } from 'react';
 
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export function useLockBodyScroll(isShowing: boolean) {
   useEffect(() => {
     if (!isShowing) {