+++ packages/web/src/app/core/services/cms/cms.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
-import * as ContentConfig from '../../../../../contentful/content.config';
 import { HttpClient } from '@angular/common/http';
 import { CMSTransformService } from './cms-transform.service';
 import { Locale } from '../localization.service';
+import { Observable, Subject } from 'rxjs';
+import { Router } from '@angular/router';
 
 @Injectable({
   providedIn: 'root',
@@ -10,52 +11,101 @@ import { Locale } from '../localization.service';
 export class CMSService {
   private entries = {};
   private entriesToSync = [];
+  private subjectAnchorsNew = new Subject<any>();
 
-  constructor(private http: HttpClient, private cmsTransformService: CMSTransformService) {}
+  constructor(
+    private http: HttpClient,
+    private cmsTransformService: CMSTransformService,
+    private router: Router,
+  ) {}
 
-  getDocumentationPageByEntryId(entryId: string, localization: Locale): Promise<any> {
+  listenContentLoadedFromCMS(): Observable<any> {
+    return this.subjectAnchorsNew.asObservable();
+  }
+
+  contentLoadedFromCMS(): void {
+    this.subjectAnchorsNew.next();
+  }
+
+  async getPageSysId(locale: number): Promise<string> {
+    const urlFull = this.router.url.split('#')[0];
+    const urlWithoutAnchor = urlFull.split('/');
+    let pageId = '';
+    await this.getMenu(locale).then((menu) => {
+      const localeKey = Object.keys(menu['pages'][0].entry.fields.pages)[locale];
+      const subMenu = menu['pages'].find((subMenu) => subMenu.path === urlWithoutAnchor[1]);
+      if (!subMenu) {
+        console.error('FOUND NO SUBMENU WITH THAT PATH');
+      }
+      if (!urlWithoutAnchor[2] && subMenu.entry.fields.landingPage) {
+        pageId = subMenu.entry.fields.landingPage[localeKey].sys.id;
+      } else if (subMenu.entry.fields.pages) {
+        const docPage = subMenu.entry.fields.pages[localeKey].find(
+          (page) => page.fields.path[localeKey] === urlWithoutAnchor[2],
+        );
+        if (!docPage) {
+          console.error('FOUND NO PAGE WITH THAT PATH');
+        }
+        pageId = docPage.sys.id;
+      }
+    });
+    return pageId;
+  }
+
+  async getDocumentationPageByEntryId(entryId: string, localization: Locale): Promise<any> {
     let locale = 'en-GB';
     if (localization === Locale['nb-NO']) {
       locale = 'nb-NO';
     }
 
-    return this.getEntry(entryId).then((data) => {
-      let pageDescription = '';
-      if (data.fields.pageDescription && data.fields.pageDescription[locale]) {
-        pageDescription = data.fields.pageDescription[locale];
-      }
-      return {
-        title: data.fields.title[locale],
-        pageDescription: pageDescription,
-        content: this.cmsTransformService.getHTML(data, locale, this.entries),
-      };
-    });
-  }
-  getDocumentationPage(pageName: string, localization: Locale): Promise<any> {
-    let locale = 'en-GB';
-    if (localization === Locale['nb-NO']) {
-      locale = 'nb-NO';
+    const data = await this.getEntry(entryId);
+    let subMenuRoute = '';
+    if (this.router.url.split('/')[2]) {
+      subMenuRoute = this.router.url.split('/')[1] + '/';
     }
-    const contentMetadata = ContentConfig[pageName];
-
-    if (!contentMetadata) {
-      console.error('ERROR: getContent - No content found for that string');
-    }
-
-    return this.getEntry(contentMetadata.contentful.entry_id).then((data) => {
-      let pageDescription = '';
-      if (data.fields.pageDescription && data.fields.pageDescription[locale]) {
-        pageDescription = data.fields.pageDescription[locale];
-      }
-      return {
-        title: data.fields.title[locale],
-        pageDescription: pageDescription,
-        content: this.cmsTransformService.getHTML(data, locale, this.entries),
-      };
-    });
+    const description = data.fields.pageDescription
+      ? this.cmsTransformService.getHTML(data, locale, 'pageDescription')
+      : '';
+    const content = data.fields.content ? this.cmsTransformService.getHTML(data, locale, 'content') : '';
+    const figmaUrl = data.fields.figmaUrl ? data.fields.figmaUrl[locale] : '';
+    const isMainPage = data.fields.isMainPage ? data.fields.isMainPage : '';
+    return {
+      title: data.fields.title[locale],
+      pageDescription: description,
+      figmaUrl: figmaUrl,
+      content: content,
+      isMainPage: isMainPage,
+      docUrl: data.fields.path[locale],
+      fullPath: subMenuRoute + data.fields.path[locale],
+    };
   }
 
-  async getMenu(localization: Locale) {
+  async getSubMenuList(localization: Locale): Promise<any> {
+    const mainMenu = await this.getMenu(localization);
+    const subMenuRoute = this.router.url.split('/')[1];
+    const subMenuList = [];
+    mainMenu.pages.forEach((element) => {
+      if (element.path === subMenuRoute) {
+        if (element.entry.fields.pages === undefined || element.entry.fields.pages === null) {
+          return;
+        }
+        const localeKey = Object.keys(element.entry.fields.pages)[localization];
+        const cmsPages = element.entry.fields.pages[localeKey];
+        cmsPages.forEach((element) => {
+          const navbarItem = {
+            title: element.fields.title[localeKey],
+            isMainPage: element.fields.isMainPage,
+            docUrl: element.fields.path[localeKey],
+            fullPath: subMenuRoute + element.fields.path[localeKey],
+          };
+          subMenuList.push(navbarItem);
+        });
+      }
+    });
+    return subMenuList;
+  }
+
+  async getMenu(localization: Locale): Promise<any> {
     let locale = 'en-GB';
     if (localization === Locale['nb-NO']) {
       locale = 'nb-NO';
@@ -78,14 +128,14 @@ export class CMSService {
     return menu;
   }
 
-  async syncEntries() {
+  async syncEntries(): Promise<any> {
     while (this.entriesToSync.length > 0) {
       const id = this.entriesToSync.pop();
       await this.getEntry(id);
     }
   }
 
-  async findEntriesWithinNode(node) {
+  async findEntriesWithinNode(node: Record<string, any>): Promise<any> {
     if (node.id && node.type === 'Entry') {
       if (!this.entries[node.id]) {
         this.entriesToSync.push(node.id);