+++ packages/web/src/app/doc-pages/components/breadcrumb-doc/breadcrumb-doc.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { getComponent } from 'src/app/shared/e-items';
+import { getComponent } from 'src/app/shared/doc-pages';
 import { breadcrumbData } from './breadcrumb-data';
 
 @Component({
@@ -41,13 +41,13 @@ export class BreadcrumbDocComponent {
     'Should not replace the primary navigation',
   ];
 
-  addBreadcrumb = () => {
+  addBreadcrumb = (): void => {
     if (this.breadcrumbsTest.length < 5) {
       this.breadcrumbsTest = this.breadcrumbsTest.concat(this.exampleBreadcrumb);
     }
   };
 
-  removeBreadcrumb = () => {
+  removeBreadcrumb = (): void => {
     if (this.breadcrumbsTest.length > 2) {
       this.breadcrumbsTest = this.breadcrumbsTest.slice(0, this.breadcrumbsTest.length - 1);
     }