+++ packages/web/src/app/doc-pages/components/carousel-doc/carousel-doc.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { getComponent } from 'src/app/shared/e-items';
+import { getComponent } from 'src/app/shared/doc-pages';
 import { carouselData } from './carousel-data';
 import { exampleContents } from 'src/app/shared/example-contents';