+++ packages/web/src/app/doc-pages/components/components.module.ts
@@ -3,7 +3,6 @@ import { LabelDocModule } from './label-doc/label-doc.module';
 import { ButtonDocModule } from './button-doc/button-doc.module';
 import { CardDocModule } from './card-doc/card-doc.module';
 import { CarouselDocModule } from './carousel-doc/carousel-doc.module';
-import { ComponentsStartModule } from './components-start/components-start.module';
 import { FormComponentsModule } from './forms/formComponents.module';
 import { LinkDocModule } from './link-doc/link-doc.module';
 import { ListDocModule } from './list-doc/list-doc.module';
@@ -42,7 +41,6 @@ import { BoxDocModule } from './box-doc/box-doc.module';
     CardDocModule,
     CarouselDocModule,
     ChipDocModule,
-    ComponentsStartModule,
     ContentLoaderDocModule,
     DividerDocModule,
     DropdownDocModule,