+++ packages/web/src/app/doc-pages/identity/spacing-doc/spacing.interface.ts
@@ -1,5 +1,0 @@
-export interface Spacing {
-  title: string;
-  size: string;
-  className?: string;
-}