+++ packages/web/src/app/shared/doc-pages.interface.ts
@@ -1,0 +1,13 @@
+export interface DocPage {
+  title: string;
+  description?: string;
+  imageUrl?: string;
+  absolutePath?: string;
+  fragmentPath?: string;
+  docUrl?: string;
+  externalUrl?: string;
+  figmaUrl?: string;
+  status?: string;
+  imageUrlOn?: string;
+  type?: string;
+}