+++ packages/web/src/app/shared/item-status.enum.ts
@@ -1,7 +1,0 @@
-export enum ItemStatus {
-  Done = 'Done',
-  Coming = 'Coming',
-  New = 'New',
-  Ignore = 'Ignore',
-  Progress = 'In progress',
-}